techtrek/requirements.txt
aniso8601==8.1.0
click==7.1.2
Flask==2.0.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
itsdangerous==2.0
Jinja2==3.0
MarkupSafe==2.0.0
pytz==2020.5
six==1.15.0
Werkzeug==2.0
python-dotenv==0.15.0
mysql-connector-python==8.0.26
